refactor: split Enrich widget code
